+++ comic_dl/honcho.py
@@ -26,6 +26,7 @@ from .sites import readComicsIO
 from .sites import japscan
 from .sites import manganelo
 from .sites import webtoons
+from .sites import lectortmo
 
 
 class Honcho(object):
@@ -93,7 +94,8 @@ class Honcho(object):
                                 chapter_range=chapter_range, conversion=kwargs.get("conversion"),
                                 keep_files=kwargs.get("keep_files"))
             return 0
-        elif domain in ["www.readcomiconline.li", "readcomiconline.li", "www.readcomicsonline.ru", "readcomicsonline.ru"]:
+        elif domain in ["www.readcomiconline.li", "readcomiconline.li", "www.readcomicsonline.ru",
+                        "readcomicsonline.ru"]:
             readcomicOnlineli.ReadComicOnlineLi(manga_url=comic_url, logger=logging,
                                                 current_directory=current_directory, sorting_order=sorting,
                                                 log_flag=log_flag, download_directory=download_directory,
@@ -237,9 +239,15 @@ class Honcho(object):
             return 0
         elif domain in ["www.webtoons.com", "webtoons.com"]:
             webtoons.Webtoons(manga_url=comic_url, logger=logging, current_directory=current_directory,
+                              sorting_order=sorting, log_flag=log_flag, download_directory=download_directory,
+                              chapter_range=chapter_range, conversion=kwargs.get("conversion"),
+                              keep_files=kwargs.get("keep_files"), image_quality=kwargs.get("image_quality"))
+            return 0
+        elif domain in ["www.lectortmo.com", "lectortmo.com"]:
+            lectortmo.LectorTmo(manga_url=comic_url, logger=logging, current_directory=current_directory,
                                 sorting_order=sorting, log_flag=log_flag, download_directory=download_directory,
                                 chapter_range=chapter_range, conversion=kwargs.get("conversion"),
-                                keep_files=kwargs.get("keep_files"), image_quality=kwargs.get("image_quality"))
-            return 0
+                                keep_files=kwargs.get("keep_files"),
+                                print_index=print_index)
         else:
             print("%s is not supported at the moment. You can request it on the Github repository." % domain)