+++ comic_dl/sites/lectortmo.py
@@ -1,13 +1,19 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+import random
+import sys
+
+from bs4 import BeautifulSoup
 
 from comic_dl import globalFunctions
 import re
 import os
 import logging
+import json
+import time
 
 
-class ComicNaver(object):
+class LectorTmo(object):
     def __init__(self, manga_url, download_directory, chapter_range, **kwargs):
 
         current_directory = kwargs.get("current_directory")
@@ -15,74 +21,76 @@ class ComicNaver(object):
         keep_files = kwargs.get("keep_files")
         self.logging = kwargs.get("log_flag")
         self.sorting = kwargs.get("sorting_order")
-        self.comic_name = self.name_cleaner(manga_url)
+        self.comic_name = None
         self.print_index = kwargs.get("print_index")
-        if "/list?" in manga_url or "list.nhn" in manga_url:
+        if "/library/" in manga_url:
             self.full_series(manga_url, self.comic_name, self.sorting, download_directory, chapter_range=chapter_range,
                              conversion=conversion, keep_files=keep_files)
-
-        elif "/detail?" in manga_url or "detail.nhn" in manga_url:
+        # https://lectortmo.com/view_uploads/979773
+        elif "/viewer/" in manga_url or "/paginated/" in manga_url or "/view_uploads/" in manga_url:
             self.single_chapter(manga_url, self.comic_name, download_directory, conversion=conversion,
                                 keep_files=keep_files)
 
-    def name_cleaner(self, url):
-        manga_name = re.search(r"titleId=(\d+)", str(url)).group(1)
-
-        return manga_name
-
     def single_chapter(self, comic_url, comic_name, download_directory, conversion, keep_files):
-        # http: // comic.naver.com / webtoon / detail.nhn?titleId = 654817 & no = 100 & weekday = tue
-        chapter_number = re.search(r"no=(\d+)", str(comic_url)).group(1)
+        comic_url = str(comic_url)
+        # https://lectortmo.com/viewer/004b1c38ce59f14291118de9f59bed7e/paginated/1
+        # https://lectortmo.com/view_uploads/979773
+        chapter_number = comic_url.split('/')[-1] if "/view_uploads/" in comic_url else comic_url.split('/')[-3]
 
         source, cookies = globalFunctions.GlobalFunctions().page_downloader(manga_url=comic_url)
-        # https://image-comic.pstatic.net/webtoon/183559/399/20180810173548_ffbf217190f59dc04bd6fc538e11d64b_IMAG01_1.jpg
-        img_regex = r'https?://(?:imgcomic\.naver\.net|image-comic\.pstatic\.net)/webtoon/\d+/\d+/.+?\.(?:jpg|png|gif|bmp|JPG|PNG|GIF|BMP)'
-        image_list = list(re.findall(img_regex, str(source)))
-        if len(image_list) == 0:
-            all_image_tags = source.find_all("img", {"alt": "comic content"})
-            for img_tag in all_image_tags:
-                image_list.append(img_tag['src'])
-        logging.debug("Image List : %s" % image_list)
+        ld_json_content = source.find_all("script", {"type": "application/ld+json"})
+        if len(ld_json_content) > 0:
+            cleaned_json_string = ld_json_content[0].next.strip().replace('\n', '')
+            loaded_json = json.loads(cleaned_json_string)
+            if loaded_json:
+                self.comic_name = comic_name = loaded_json['headline']
+        links = []
+        file_names = []
+        img_url = self.extract_image_link_from_html(source=source)
+        links.append(img_url)
+        img_extension = str(img_url).rsplit('.', 1)[-1]
+        unique_id = str(img_url).split('/')[-2]
+        file_names.append('{0}.{1}'.format(1, img_extension))
 
-        file_directory = globalFunctions.GlobalFunctions().create_file_directory(chapter_number, comic_name)
-        # directory_path = os.path.realpath(file_directory)
+        total_page_list = source.find("select", {"id": "viewer-pages-select"})
+        last_page_number = 0
+        options = total_page_list.findAll('option')
+        if len(options) > 0:
+            last_page_number = int(options[-1]['value'])
+        if last_page_number <= 0:
+            print("Couldn't find all the pages. Exiting.")
+            sys.exit(1)
+        for page_number in range(2, last_page_number):
+            current_url = "https://lectortmo.com/viewer/{0}/paginated/{1}".format(unique_id, page_number)
+            print("Grabbing details for: {0}".format(current_url))
+            source, cookies = globalFunctions.GlobalFunctions().page_downloader(manga_url=current_url, cookies=cookies)
+            image_url = self.extract_image_link_from_html(source=source)
+            links.append(image_url)
+            img_extension = str(image_url).rsplit('.', 1)[-1]
+            file_names.append('{0}.{1}'.format(page_number, img_extension))
+            time.sleep(random.randint(1, 6))
+        file_directory = globalFunctions.GlobalFunctions().create_file_directory(chapter_number, self.comic_name)
+
         directory_path = os.path.realpath(str(download_directory) + "/" + str(file_directory))
-        print("Directory Path : %s" % str(directory_path))
 
         if not os.path.exists(directory_path):
             os.makedirs(directory_path)
-
-        links = []
-        file_names = []
-        for current_chapter, image_link in enumerate(image_list):
-            current_chapter += 1
-
-            # Fix for #18
-            file_name = str(globalFunctions.GlobalFunctions().prepend_zeroes(current_chapter, len(image_list))) + ".jpg"
-
-            file_names.append(file_name)
-            links.append(image_link)
-
-        globalFunctions.GlobalFunctions().multithread_download(chapter_number, comic_name, comic_url, directory_path,
+        globalFunctions.GlobalFunctions().multithread_download(chapter_number, self.comic_name, comic_url, directory_path,
                                                                file_names, links, self.logging)
-            
-        globalFunctions.GlobalFunctions().conversion(directory_path, conversion, keep_files, comic_name,
+
+        globalFunctions.GlobalFunctions().conversion(directory_path, conversion, keep_files, self.comic_name,
                                                      chapter_number)
 
         return 0
 
     def full_series(self, comic_url, comic_name, sorting, download_directory, chapter_range, conversion, keep_files):
-        comic_type = re.findall(r'comic.naver.com/(.*?)/', str(comic_url))[0]
         source, cookies = globalFunctions.GlobalFunctions().page_downloader(manga_url=comic_url)
-        # print(source)
-
-        latest_chapter = re.findall(r"no=(\d+)[\&|\"]", str(source))[1]
 
         all_links = []
+        all_chapter_links = source.find_all("a", {"class": "btn btn-default btn-sm"})
+        for chapter in all_chapter_links:
+            all_links.append(chapter['href'])
 
-        for x in range(1, int(latest_chapter) + 1):
-            chapter_url = "http://comic.naver.com/{0}/detail.nhn?titleId={1}&no={2}".format(comic_type, comic_name, x)
-            all_links.append(chapter_url)
         logging.debug("All Links : %s" % all_links)
 
         # Uh, so the logic is that remove all the unnecessary chapters beforehand
@@ -120,7 +128,8 @@ class ComicNaver(object):
                     break  # break to continue processing other mangas
                 # if chapter range contains "__EnD__" write new value to config.json
                 # @Chr1st-oo - modified condition due to some changes on automatic download and config.
-                if chapter_range != "All" and (chapter_range.split("-")[1] == "__EnD__" or len(chapter_range.split("-")) == 3):
+                if chapter_range != "All" and (
+                        chapter_range.split("-")[1] == "__EnD__" or len(chapter_range.split("-")) == 3):
                     globalFunctions.GlobalFunctions().addOne(comic_url)
         elif str(sorting).lower() in ['old', 'asc', 'ascending', 'oldest', 'a']:
             # print("Running this")
@@ -134,7 +143,15 @@ class ComicNaver(object):
                     break  # break to continue processing other mangas
                 # if chapter range contains "__EnD__" write new value to config.json
                 # @Chr1st-oo - modified condition due to some changes on automatic download and config.
-                if chapter_range != "All" and (chapter_range.split("-")[1] == "__EnD__" or len(chapter_range.split("-")) == 3):
+                if chapter_range != "All" and (
+                        chapter_range.split("-")[1] == "__EnD__" or len(chapter_range.split("-")) == 3):
                     globalFunctions.GlobalFunctions().addOne(comic_url)
 
         return 0
+
+    def extract_image_link_from_html(self, source):
+        image_tags = source.find_all("img", {"class": "viewer-image viewer-page"})
+        img_link = None
+        for element in image_tags:
+            img_link = element['src']
+        return img_link